Work around clang's warning about using jlong

+++ src/hotspot/share/runtime/fieldDescriptor.cpp
@@ -212,21 +212,21 @@ void fieldDescriptor::print_on_for(outputStream* st, oop obj) {
     if (UseCompressedOops) {
       st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));
     } else {
-      st->print(" (" INT64_FORMAT_X_0 ")", obj->long_field(offset()));
+      st->print(" (" INT64_FORMAT_X_0 ")", (int64_t)obj->long_field(offset()));
     }
 #else
     st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));
 #endif
   } else { // Primitives
     switch (ft) {
-      case T_LONG:    st->print(" (" INT64_FORMAT_X_0 ")", obj->long_field(offset()));  break;
-      case T_DOUBLE:  st->print(" (" INT64_FORMAT_X_0 ")", obj->long_field(offset()));  break;
-      case T_BYTE:    st->print(" (" INT8_FORMAT_X_0  ")", obj->byte_field(offset()));  break;
-      case T_CHAR:    st->print(" (" INT16_FORMAT_X_0 ")", obj->char_field(offset()));  break;
-      case T_FLOAT:   st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));   break;
-      case T_INT:     st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));   break;
-      case T_SHORT:   st->print(" (" INT16_FORMAT_X_0 ")", obj->short_field(offset())); break;
-      case T_BOOLEAN: st->print(" (" INT8_FORMAT_X_0  ")", obj->bool_field(offset()));  break;
+      case T_LONG:    st->print(" (" INT64_FORMAT_X_0 ")", (int64_t)obj->long_field(offset())); break;
+      case T_DOUBLE:  st->print(" (" INT64_FORMAT_X_0 ")", (int64_t)obj->long_field(offset())); break;
+      case T_BYTE:    st->print(" (" INT8_FORMAT_X_0  ")", obj->byte_field(offset()));          break;
+      case T_CHAR:    st->print(" (" INT16_FORMAT_X_0 ")", obj->char_field(offset()));          break;
+      case T_FLOAT:   st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));           break;
+      case T_INT:     st->print(" (" INT32_FORMAT_X_0 ")", obj->int_field(offset()));           break;
+      case T_SHORT:   st->print(" (" INT16_FORMAT_X_0 ")", obj->short_field(offset()));         break;
+      case T_BOOLEAN: st->print(" (" INT8_FORMAT_X_0  ")", obj->bool_field(offset()));          break;
     default:
       ShouldNotReachHere();
       break;
